M2_T1-6

+++ tehtävät/OH1_M2_T2.py
@@ -1,6 +1,6 @@
 import math
 
-sade = int(input('säde: '))
+sade = float(input('säde: '))
 
 pintaala = math.pi*sade*sade
 print(pintaala)
