+++ tehtävät/OH1_M2_T3.py
@@ -1,5 +1,5 @@
-korkeus = int(input('korkeus: '))
-leveys = int(input('leveys: '))
+korkeus = float(input('korkeus: '))
+leveys = float(input('leveys: '))
 
 pintaala = (korkeus*leveys)
 piiri = 2*korkeus+2*leveys
