+++ tehtävät/OH1_M2_T4.py
@@ -1,6 +1,6 @@
-luku_1 = int(input("Luku 1: "))
-luku_2 = int(input("Luku 2: "))
-luku_3 = int(input("Luku 3: "))
+luku_1 = float(input("Luku 1: "))
+luku_2 = float(input("Luku 2: "))
+luku_3 = float(input("Luku 3: "))
 
 summa = luku_1 + luku_2 + luku_3
 tulo = luku_1*luku_2*luku_3
